Update book_graph.py
Implemented a method to filter out unhelpful tags (fiction, non-fiction), and force the inclusion of a country-related tag. This tag can either be the setting or the author's country of origin, which is usually a useful characteristic for narrowing down book choices. I may end up including both the setting and author's country of origin eventually.